+++ web/src/main/java/com/github/adamovichas/project/web/servlet/BetServlet.java
@@ -54,7 +54,7 @@ public class BetServlet extends HttpServlet {
     @Override
     protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
         Long eventId = Long.valueOf(req.getParameter("eventId"));
-        int moneyForBet = Integer.parseInt(req.getParameter("money_for_bet"));
+        Double moneyForBet = Double.parseDouble(req.getParameter("money_for_bet"));
         EventDTO eventDTO = dataEvent.getEventById(eventId);
         String factorName = req.getParameter("factorName");
         FactorDTO factorDTO = util.getFactorByName(eventDTO, factorName);
